requirements.txt
streamlit
openai
python-dotenv
streamlit-autorefresh